backend/pyproject.toml
[tool.poetry]
name = "wareflow-backend"
version = "0.1.0"
description = ""
authors = ["hiteshshimpi <shimpihitesh2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
gunicorn = "^22.0.0"
psycopg2-binary = "^2.9.9"
python-jose = "^3.3.0"
python-dotenv = "^1.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = "^2.7.4"
pydantic-settings = "^2.3.3"
fastapi-sqlalchemy = "^0.2.1"
alembic = "^1.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
